refactored and split repositories - fixed env test according to pylint - added makefile default target - sorted requirements
