+++ app/presentation/commands/runserver/__init__.py
@@ -1,7 +1,6 @@
 import structlog
 
-from app.data import Storage
-from app.data.repository import DataRepository
+from app.data import Storage, repositories
 from app.domain import usecases
 from app.presentation import server
 from app.presentation.commands.runserver.config import parse_config
@@ -17,11 +16,16 @@ def start(config_path: str):
         ],
     )
 
+    logger = structlog.get_logger()
+
     storage = Storage(cfg.storage)
     storage.connect()
 
-    repo = DataRepository(storage)
-    actions = usecases.Actions(repo)
+    common_repo = repositories.CommonRepository(storage, logger)
+    layer0_repo = repositories.Layer0Repository(storage, logger)
+    layer1_repo = repositories.Layer1Repository(storage, logger)
+
+    actions = usecases.Actions(common_repo, layer0_repo, layer1_repo, logger)
 
     server.start(cfg.server, actions)
     storage.disconnect()
