+++ app/presentation/server/handlers/ping.py
@@ -1,3 +1,4 @@
+from dataclasses import dataclass
 from typing import Any
 
 from aiohttp import web
@@ -6,8 +7,11 @@ from aiohttp_apispec import docs
 from app import domain
 
 
+@dataclass
+class PingResponse:
+    ping: str = "pong"
+
+
 @docs(summary="Test that service is up and running")
 async def ping(_: domain.Actions, __: web.Request) -> Any:
-    return {
-        "ping": "pong",
-    }
+    return PingResponse()
