+++ requirements.txt
@@ -1,22 +1,23 @@
-astropy
-astroquery
-pandas
 aiohttp
 aiohttp_apispec
+astropy
+astroquery
+black
+click
+isort
+jinja2
 marshmallow
+mypy
+pandas
+pandas-stubs
+parameterized
+psycopg[binary,pool]
+pylint
+pytest
+pyyaml
 requests
 setuptools
-pyyaml
-click
-black
-isort
-psycopg[binary,pool]
 structlog
 testcontainers-postgres
-jinja2
-pandas-stubs
-types-pyyaml
 types-psycopg2
-pytest
-pylint
-parameterized
+types-pyyaml
