+++ tests/integration/sources_test.py
@@ -1,6 +1,8 @@
 import unittest
 
-from app.data import repository
+import structlog
+
+from app.data import repositories
 from app.domain import model, usecases
 from app.lib import testing
 
@@ -11,8 +13,10 @@ class SourcesTest(unittest.TestCase):
         cls.storage = testing.TestPostgresStorage("postgres/migrations")
         cls.storage.start()
 
-        cls.repo = repository.DataRepository(cls.storage.get_storage())
-        cls.actions = usecases.Actions(cls.repo)
+        common_repo = repositories.CommonRepository(cls.storage.get_storage(), structlog.get_logger())
+        layer0_repo = repositories.Layer0Repository(cls.storage.get_storage(), structlog.get_logger())
+        layer1_repo = repositories.Layer1Repository(cls.storage.get_storage(), structlog.get_logger())
+        cls.actions = usecases.Actions(common_repo, layer0_repo, layer1_repo, structlog.get_logger())
 
     @classmethod
     def tearDownClass(cls):
